Add simple sample files for react native lib and rename reactNative directories
