+++ packages/nx-plugin/src/generators/app/index.ts
@@ -1,3 +1,3 @@
 export * from "./node";
 export * from "./react";
-export * from "./react-native";
+export * from "./reactNative";
