+++ packages/nx-plugin/src/generators/app/reactNative/index.ts
