+++ packages/nx-plugin/src/generators/app/reactNative/reactNativeAppGenerator.ts
