+++ packages/nx-plugin/src/generators/app/reactNative/schema.json
