+++ packages/nx-plugin/src/generators/lib/reactNative/index.ts
