+++ packages/nx-plugin/src/generators/lib/reactNative/schema.json
@@ -21,7 +21,5 @@
       "default": false
     }
   },
-  "required": [
-    "name"
-  ]
+  "required": ["name"]
 }
