+++ packages/nx-plugin/src/generators/project/index.ts
@@ -6,7 +6,7 @@ export * from "./project";
 export * from "./projectGeneratorBase.schema";
 export * from "./projectJson";
 export * from "./react";
-export * from "./react-native";
+export * from "./reactNative";
 export * from "./readme";
 export * from "./samples";
 export * from "./tsconfig";
