+++ packages/nx-plugin/src/generators/project/reactNative/index.ts
