+++ packages/nx-plugin/src/generators/project/reactNative/reactNativeProjectGenerator.ts
