+++ packages/nx-plugin/src/generators/project/reactNative/reactNativeProjectGeneratorSchema.ts
