+++ packages/nx-plugin/src/generators/project/samples/index.ts
@@ -1,2 +1,3 @@
 export * from "./node";
+export * from "./reactNative";
 export * from "./ts";
