+++ packages/nx-plugin/src/generators/project/samples/reactNative/index.ts
@@ -1,0 +1,1 @@
+export * from "./lib";
