+++ packages/nx-plugin/src/generators/project/samples/reactNative/lib/files/src/index.ts__template__
@@ -1,0 +1,1 @@
+export * from "./HelloWorld";
