+++ packages/nx-plugin/src/generators/project/samples/reactNative/lib/index.ts
@@ -1,0 +1,1 @@
+export * from "./copyReactNativeLibSample";
